Add a menu option for showjoinaddress; disable it by default

See merge request STJr/SRB2!1115

+++ src/m_menu.h
@@ -327,7 +327,7 @@ typedef struct menuitem_s
 	void *itemaction;
 
 	// hotkey in menu or y of the item
-	UINT8 alphaKey;
+	UINT16 alphaKey;
 } menuitem_t;
 
 extern menuitem_t MP_RoomMenu[];
